requirements.txt
g4f==0.1.9.4
python-dotenv==0.19.0
python-telegram-bot==13.7
